Trap the interrupt instead of rescuing.
Using `rescue` to capture the interrupt is a bit strange. While it could
be useful in certain circumstances, we are only using it as an exit
hook.

We do not use `at_exit` here as we want to treat the interrupt as
"normal" instead of it being an exception. Using `at_exit` would display
the error at the end of the script run. Trapping it lets us decide what
to do, Ruby will not treat it as an error.