Wait until the beacon is removed
This adds a wait after a flash until the beacon is removed. This helps
people who get distracted or may have missed hearing the tone notifying
the end of the flashing process. Without this wait, when the person
checks the output, there is no way for them to know if what they are
seeing is for the beacon plugged in or the previous beacon.