Check file and flash DFU
Checks to see if firmware file is of the correct size, this will help to prevent flashing incompatible firmware files which could corrupt the device

Also, will continue flashing if the dongle was stuck in DFU mode. Previously, using the `-w` option, if a dongle that was already in DFU mode was inserted, it would not be flashed. However, it is safe (and probably desirable) to flash dongles that are in DFU mode.